+++ R/server.R
@@ -505,7 +505,7 @@ server <- function(input, output, session) {
   })
   
   observeEvent(input$solute_select_sp, {
-    updateSelectInput(session, "solute_select_ts", selected = input$solute_select_sp )
+    #updateSelectInput(session, "solute_select_ts", selected = input$solute_select_sp )
     tr<-as.numeric(csite$ui_attr$plume_thresh[as.character(input$solute_select_sp)])
     updateNumericInput(session,"plume_thresh_pd",value=tr)
     
@@ -559,7 +559,14 @@ server <- function(input, output, session) {
     csite$ui_attr$ts_options[input$ts_true_options] <<- TRUE
     
     
-    plotTimeSeries(csite, input$solute_select_ts, input$sample_loc_select_ts, input$check_threshold)
+    #plotTimeSeries(csite, input$solute_select_ts, input$sample_loc_select_ts, input$check_threshold)
+    plotTimeSeries(
+      csite=csite,
+      substance=input$solute_select_ts,
+      location=input$sample_loc_select_ts,
+      show_thresh=input$check_threshold,
+      timepoint=input$timepoint_ts_idx
+    )
     
   })
 
@@ -884,16 +891,26 @@ server <- function(input, output, session) {
     # Re-Calculate groundwater flows (depends on aggregation date).
     csite$GW.Flows <<- evalGWFlow(csite$All.Data$Agg_GW_Data)
     
-        
     # Update UI time points of slider.
     dates_tmp <- format(csite$All.Data$All_Agg_Dates, "%d-%m-%Y")
     csite$ui_attr$timepoints   <<- dates_tmp
-    
+
     # Set new time point to last date.
     new_timepoint_idx <- length(dates_tmp)
     
     # Update slider inputs: Spatial plot and in Trend table.
     outp <- pasteAggLimit(csite$ui_attr$timepoints[new_timepoint_idx], csite$GWSDAT_Options$Aggby)
+    
+    
+    if(new_timepoint_idx==input$timepoint_sp_idx){ 
+      
+    #check for event when new length time points is unchanged. Manually change to something else - so update of slider label is performed.
+    updateSliderInput(session, "timepoint_sp_idx", value = max(1,new_timepoint_idx-1),
+                      min = 1, max = length(csite$ui_attr$timepoints), label ="", step = 1)
+    updateSliderInput(session, "timepoint_tt_idx", value = max(1,new_timepoint_idx-1),
+                      min = 1, max = length(csite$ui_attr$timepoints), label = "", step = 1) 
+  
+    }
     
     updateSliderInput(session, "timepoint_sp_idx", value = new_timepoint_idx,
                       #min = 1, max = length(csite$ui_attr$timepoints), label = paste0("Time: ", outp), step = 1)
@@ -999,7 +1016,14 @@ server <- function(input, output, session) {
   })
     
   
-  
+  # 
+  # ### adding Disclaimer in the plottrendtable
+  # 
+  # output$note <- renderText({
+  #   "Disclaimer : The slider displays only the dates which are available in the data."
+  # })
+  # 
+  # # 
   
   output$trend_table <- renderUI({
     
@@ -1018,8 +1042,12 @@ server <- function(input, output, session) {
       return(NULL)
     }
 
-    plotTrendTable(csite, as.Date(csite$ui_attr$timepoints[input$timepoint_tt_idx], "%d-%m-%Y"),
-               input$trend_or_threshold, input$color_select_tt)
+    plotTrendTable(csite, 
+                   as.Date(csite$ui_attr$timepoints[input$timepoint_tt_idx], "%d-%m-%Y"),
+                #input$timepoint_tt_idx,
+               input$trend_or_threshold, 
+               input$color_select_tt,
+               input$substance_select_tt)
   })
 
   
@@ -1059,6 +1087,28 @@ server <- function(input, output, session) {
     
   })
   
+  
+
+  
+  observeEvent(input$selectall_stp, {
+
+    if (input$selectall_stp %% 2 == 0){
+      updateSelectInput(session, "sample_loc_select_stp",selected=csite$ui_attr$sample_loc_names)
+    }else{
+      updateSelectInput(session, "sample_loc_select_stp",selected="")
+    }
+
+    
+  })
+  
+  
+  
+  
+  
+  
+  
+  
+  
   updateNAPL <- function(location, substance) {
     
     tmp_napl <- existsNAPL(csite$All.Data, location, substance) 
@@ -1085,10 +1135,8 @@ server <- function(input, output, session) {
   # to Spatial Plot panel.
   observeEvent({input$solute_select_ts; 
     input$sample_loc_select_ts}, {
-    
     updateNAPL(input$sample_loc_select_ts, input$solute_select_ts)  
-  
-    updateSelectInput(session, "solute_select_sp", selected = input$solute_select_ts ) 
+    #updateSelectInput(session, "solute_select_sp", selected = input$solute_select_ts[length(input$solute_select_ts)] ) 
   })
   
   #observeEvent(input$solute_select_sp, {
@@ -1175,7 +1223,7 @@ server <- function(input, output, session) {
       
       if (input$export_format_ts == "pptx") {
         
-        makeTimeSeriesPPT(csite=csite, fileout=file, substance=input$solute_select_ts, location=input$sample_loc_select_ts,show_thresh=input$check_threshold,
+        makeTimeSeriesPPT(csite=csite, fileout=file, substance=input$solute_select_ts, location=input$sample_loc_select_ts,show_thresh=input$check_threshold,timepoint=input$timepoint_ts_idx,
                           width  = input$img_width_px, height = input$img_height_px)
         
       } 
@@ -1187,7 +1235,7 @@ server <- function(input, output, session) {
         if (input$export_format_ts == "jpg") jpeg(file, width = input$img_width_px, height = input$img_height_px, quality = input$img_jpg_quality) 
         if (input$export_format_ts == "wmf") grDevices::win.metafile(file, width = input$img_width_px/100, height = input$img_height_px/100)
         
-        plotTimeSeries(csite, substance=input$solute_select_ts, location=input$sample_loc_select_ts,show_thresh=input$check_threshold)
+        plotTimeSeries(csite=csite, substance=input$solute_select_ts, location=input$sample_loc_select_ts,show_thresh=input$check_threshold,timepoint=input$timepoint_ts_idx)
         dev.off()
       }
     }
@@ -1228,8 +1276,11 @@ server <- function(input, output, session) {
   })
   
   observeEvent(input$export_format_sp , {
-    if (input$export_format_sp == "asc") {
+    if (input$export_format_sp == "asc" ) {
       shinyjs::disable("save_spatial_plot")
+    }
+    if (input$export_format_sp != "asc"){
+      shinyjs::enable("save_spatial_plot")
     }
   })
   observeEvent(input$resolution , {
@@ -1359,6 +1410,18 @@ server <- function(input, output, session) {
       
     }
   )
+  
+  
+  observeEvent(input$selectall_wr, {
+    
+    if (input$selectall_wr %% 2 == 0){
+      updateSelectInput(session, "sample_loc_select_wr",selected=csite$ui_attr$sample_loc_names)
+    }else{
+      updateSelectInput(session, "sample_loc_select_wr",selected="")
+    }
+    
+    
+  })
   
   
   output$save_plumestats_plot <- downloadHandler(
@@ -1511,7 +1574,7 @@ output$generate_timeseries_anim_ppt <- downloadHandler(
     
     content <- function(file) {
       
-      makeTimeSeriesAnimationPPT(csite=csite, fileout=file, substance=input$solute_select_ts, location=input$gwwellreportsample_loc_select_wr,Layout=input$gwwellreportlayout,show_thresh=input$check_threshold,
+      makeTimeSeriesAnimationPPT(csite=csite, fileout=file, substance=input$solute_select_ts, location=input$gwwellreportsample_loc_select_wr,Layout=input$gwwellreportlayout,show_thresh=input$check_threshold,timepoint=input$timepoint_ts_idx,
                         width  = input$img_width_px, height = input$img_height_px)
       
     }